+++ okf/__init__.py
@@ -15,7 +15,7 @@ CLI:
     okf summarize ./okf_bundle
 """
 
-__version__ = "0.1.34"
+__version__ = "0.1.35"
 __author__  = "Umair Baig"
 __license__ = "MIT"